Move loadShows function to modules
